+++ CP2013_WordOfMouth/JSON/ConverterBooking.cs
@@ -8,13 +8,13 @@ namespace CP2013_WordOfMouth.JSON
 {
     class ConverterBooking
     {
-        public int seshID { get; set; }
+        public long seshID { get; set; }
         public int appointment_type { get; set; }
         public int dentist { get; set; }
         public int day { get; set; }
         public string time { get; set; }
 
-        public ConverterBooking(int seshID, int appointment_type, int dentist, int day, string time)
+        public ConverterBooking(long seshID, int appointment_type, int dentist, int day, string time)
         {
             this.seshID = seshID;
             this.appointment_type = appointment_type;
